metering: keep provider details in SDK and relocate traffic rules
