+++ pkg/metering/config_test.go
@@ -14,11 +14,9 @@
 package metering
 
 import (
-	"encoding/json"
 	"testing"
 
 	"github.com/pingcap/metering_sdk/config"
-	"github.com/pingcap/metering_sdk/storage"
 	"github.com/stretchr/testify/require"
 )
 
@@ -30,18 +28,9 @@ func TestValidateConfig(t *testing.T) {
 	}{
 		{"disabled nil", nil, true},
 		{"disabled empty", &config.MeteringConfig{}, true},
-		{"missing localfs options", &config.MeteringConfig{Type: storage.ProviderTypeLocalFS}, false},
-		{"localfs", config.NewMeteringConfig().WithLocalFS(t.TempDir()), true},
-		{"s3", config.NewMeteringConfig().WithS3("us-west-2", "bucket"), true},
-		{"s3 missing bucket", config.NewMeteringConfig().WithS3("us-west-2", ""), false},
-		{"oss", config.NewMeteringConfig().WithOSS("region", "bucket"), true},
-		{"cos endpoint", &config.MeteringConfig{Type: storage.ProviderTypeCOS, Bucket: "bucket", Endpoint: "https://cos.example.com"}, true},
-		{"cos missing region and endpoint", &config.MeteringConfig{Type: storage.ProviderTypeCOS, Bucket: "bucket"}, false},
-		{"azure", config.NewMeteringConfig().WithAzure("account", "container"), true},
-		{"unsupported", &config.MeteringConfig{Type: "unknown"}, false},
-		{"pool traversal", config.NewMeteringConfig().WithS3("region", "bucket").WithSharedPoolID("../pool"), false},
-		{"endpoint credentials", config.NewMeteringConfig().WithS3("region", "bucket").WithEndpoint("https://user:secret@example.com"), false},
-		{"endpoint token", config.NewMeteringConfig().WithS3("region", "bucket").WithEndpoint("https://example.com?token=secret"), false},
+		{"SDK owns provider configuration", &config.MeteringConfig{Type: "s3"}, true},
+		{"pool", &config.MeteringConfig{Type: "s3", SharedPoolID: "pool1"}, true},
+		{"pool traversal", &config.MeteringConfig{Type: "s3", SharedPoolID: "../pool"}, false},
 	} {
 		t.Run(tc.name, func(t *testing.T) {
 			err := ValidateConfig(tc.cfg)
@@ -49,29 +38,7 @@ func TestValidateConfig(t *testing.T) {
 				require.NoError(t, err)
 			} else {
 				require.Error(t, err)
-				require.NotContains(t, err.Error(), "secret")
 			}
 		})
 	}
-}
-
-func TestRedactConfig(t *testing.T) {
-	c := &config.MeteringConfig{
-		Endpoint: "https://user:secret@example.com?token=secret#secret",
-		AWS:      &config.MeteringAWSConfig{AccessKey: "secret", SecretAccessKey: "secret", SessionToken: "secret"},
-		OSS:      &config.MeteringOSSConfig{AccessKey: "secret", SecretAccessKey: "secret", SessionToken: "secret"},
-		COS:      &config.MeteringCOSConfig{AccessKey: "secret", SecretAccessKey: "secret", SessionToken: "secret"},
-		Azure:    &config.MeteringAzureConfig{AccountKey: "secret", SASToken: "secret"},
-	}
-	before, err := json.Marshal(c)
-	require.NoError(t, err)
-	redacted := RedactConfig(c)
-	data, err := json.Marshal(redacted)
-	require.NoError(t, err)
-	require.NotContains(t, string(data), "secret")
-	require.Equal(t, "https://example.com", redacted.Endpoint)
-	after, err := json.Marshal(c)
-	require.NoError(t, err)
-	require.Equal(t, before, after)
-	require.Nil(t, RedactConfig(nil))
 }
